[FIX] account: auto post valid invoices
Previously, when account moves were set to auto post, and the cron was triggered to post them, if a single invoice triggered a user error, none of them were posted due to a User error with missing required field to post, archived journal, inactive currency, ...). This commit enables to post all valid invoices. The invoices that could not be posted are now visible on the accounting dashboard on their respective journal card in the 'to check' line, and a note is written on the chatter of the problematic invoices, describing the reason why the move could not be posted.

task-3244310