[IMP] project_todo: prevent deletion/archiving of reminder activity type
Before this commit, If we create a to-do from activity menu then reminder type
activity get created. if reminder activity type is archive or deleted then an
activity with no type is generated.
Prevented reminder type activity from deletion and archiving so the problem of
activity with no type is not get created.

task-3196864