[FIX] base, website: avoid invalid website_id leaf
In some test on runbot, an error can occur during the nightly

    `ValueError: Invalid field project.task.website_id in leaf ('website_id', 'in', (False, 1)`

This error was appearing in multiple tests with different models, and
this is changing with the time and the versions.

This is actually a strange interaction between tests and installed
module. As an exemple, lets take the error in test_billing_employee_rate
(https://runbot.odoo.com/web/#id=21961&view_type=form&model=runbot.build.error&menu_id=405&cids=1)

- It is'nt possible to reproduce it on an installed database only
executing test, the test must be ran at the same time as the test order
during install is not the same as the test order executed alone
(TestSaleProject executed after TestWebsiteControllers during install
and before otherwise)

- the minimal subset of modules needed was
`sale_project,test_website,website_event_sale,sale_timesheet`

Only executing `test_billing_employee_rate` is not enought, we need to
execute at least `test_02_search_controller` before.

This is a combination of overrides needed so that writting on
`project_root_id` will trigger a `_modified_triggers`, searching for a
`project_root_id`, getting a FALSELEAF and searching with it.

Note that to reproduce this issue fbd38450b1 must be checkout, since the
next commit, fb88a7448c changes some code arround project_root_id making
it impossible to reproduce with the proposed combination.

This same FALSELEAF is actually altered in `sitemap_qs2dom`, appending
the website_domain by mistake.

The proposed solution makes this leaf immutable using a tuple to force
a copy to change it in a list if it needs to be modified.