[IMP] tests: execute post install test sorted
There was a difference between executing tests in the same execution has
the install of after the execution, because updated_modules is set.

The idea is to test all modules that where updated or installed when
using --test-tags with -i or -u for "old habits".

To make post install execution consistent in the two cases, the modules
are now sorted by name in all cases.