[FIX] mail: invalid literal base10 for partner to field in email template
This problem arises when the customer enters an unsupported data type in the
partner To(partner to) field of the email template.

Step To Reproduce:-
- Install the sale_management,mail.
- Open Email Template
- Go to Note Page > Click on Email Configuration > To (Partners)
  > Enter String  > Save It.
- Open Sales > open one record > Click on Send By Email >
  (Select that template which you changed) > Send.
Trackback would be appear.

When a customer attempts to send an email to a user by filling out the Partner
To field's string information, at that time error will be generated,

sentry-4336102278