[FIX] website: solve some snippets color issues
Prior to this commit, some snippets used colors that were not changed or
were not visible (white text on white background) depending on the color
preset, which created color issues.

This commit adapts colors to make elements readable and maintain
consistency with the theme design.

Note: this could be partially backported (with stable adaptation) later.

task-3097005

Co-authored-by: qsm-odoo <qsm@odoo.com>