[FIX] base: raise error while invalid field in domain
When the user tries to access the model after entering an invalid field in a
domain in the record rule, the traceback appears.

Steps to produce:
1. install Fleet > Settings > Technical > Database Structures > Models
2. Select fleet.vehicle.log.services model from the list > In record rules, in a
new line, add an invalid domain having a field not present in that model,
now try to access the Fleet module and the traceback will be generated
Error: ValueError: Invalid field fleet.vehicle.log.services.cost_id in leaf
('cost_id.vehicle_id.driver_id', '=', 4156)

This commit adds a constraint to check whether the field is present in the model
or not, otherwise raises a validation error.

sentry-4335425435