[IMP] web: add install prompt and service This commit implements a notification suggesting the user to install the PWA of Odoo, with a button allowing to show the native installation prompt of the browser when available.
For iOS and macOS devices supporting the PWA, a custom prompt had to be
created, since those OS's don't suggest natively to install the site as
an application, but only allow it from the share options on iPhone/iPad,
and via the file menu on desktop (starting Safari 17).

But, in case Apple adds an installation prompt in future releases, the
service listen to the BeforeInstallPromptEvent anyway. In that case, the
native prompt would be called instead of the custom one.