[FIX] website_forum: color issues
=== ISSUE 1 ===
Prior to this commit, card colors in the index list were static,
creating a contrast issue if the page background color changed.

This commit fixes this issue.

=== ISSUE 2 ===
Prior to this commit, the popover shadow color was set to gray 400.
This created a color issue when the background color was dark.

To fix that and maintain consistency, we use the common Bootstrap
shadow.

task-3097005