[IMP] website: Expose manual models on the website
This commit implements a way to expose any manual model publicly on the website.
A manual model is a model that has been created by the user at runtime, via the technical menu,
via studio, or with an xml declaration.

The main interface to easily create such page is in Studio, in the tab Website Integration.

The route on which the models are exposed is:
"/model/<string:page_name_slugified>",

With the derived routes:
"/model/<string:page_name_slugified>/page/<int:page_number>",
"/model/<string:page_name_slugified>/<string:record_slug>"

Co-authored-by: Florent Dardenne (dafl) <dafl@odoo.com>

task-id-3231144