[FIX] mail: performance test updates
A recent change to add notifications on the mail composer results
in additional queries during the message sending process
for mass composers.

1 to create the notifications, and a few more for notification handling.
Such as setting the state of the notification.
This results in an increase of at least 2 querries per email sent
as we set a transient failure during processing each time
and update the notifications after each email after processing.

This updates all the relevant tests at once.

task-2199202