[FIX] clipboard: invalidate internal clipboard on alteration in cut zone
This commit introduces a method 'isInvalidatedBy' which returns true if
the content, style or borders of cut zone are altered before paste. Else,
it returns false. On the basis of this boolean, internal clipboard is
invalidated.

The method works as follows:

Content/Style:

It handles the "UPDATE_CELL" command and returns true if content or
style of clipped zone is altered, else it returns false. The alteration
is checked by comparing the content of cmd with content of clipped zone.

Borders:

It handles the "SET_ZONE_BORDERS" command and if the zone in which
border is being set intersects with clipped zone, it compares current
borders of clipped zone with those at the time of cut. It returns true
if any border of clipped zone is altered, else it returns false.

Task ID: 3358431