[IMP] web, mail: improve dark mode colors use
This commit introduces some adjustments to the dark mode color scheme.

=== Why do we introduce changes in Community ===

Some files related to the dark mode are located in the Community edition,
which means they need to be override in that specific location.

=== Why do we introduce changes on the properties_fields inputs ===

The inputs inside the property field modal were styled in a different way
than the regular input. We apply some styling on these inputs to make them
look like the regular one. This brings some consistency and also fix
contrast issues in dark mode.

task-3201038