[FIX] website: clear cache on ir_asset archive
Since #121376 a clear_cache was removed when unlinking an attachement

This clear_cache was not usefull when restarting a server with new
sources, an other operation changing the content of an assets should
invalidate the cache manually. This is the case od ir_assets CUD
operations.

Unfortunately, a manual update was left missing in website, when
archiving ir_assets used for snippets.

This was discovered on runbot, with two workers, when the first workers
generates assets before the cron, and the second one after the cron.

The second worker unlinks attachements creating an inconsistency in the
cache of the first one.

This problem can be solved quickly by invalidating the assets cache in
the cron manually but this will be done in all cases. The proposed
solution will check the ir.assets that should change and only clear the
cache if the state changed.

Regarding performances, this should actually be a slight improvement
since at the cost of one more select to prefetch the record we can avoid
multiple update, one per snippet. In most case no update at all should
be done, at most 2 can be done (one for archive, one for unarchive).

Example with some assets to unarchive:

Before:
TOTAL ENTRIES: 277
SELECT: 158 (~0.16591858863830566s)
UNKWOW: 54 (~0.0202481746673584s)
UPDATE: 65 (~0.03202557563781738s)

After:
TOTAL ENTRIES: 214
SELECT: 159 (~0.1663439826965332s)
UNKWOW: 54 (~0.029229164123535156s)
UPDATE: 1 (~0.0002865791320800781s)