[FIX] website_slides: handle invalid type for course duration
When the user tries to change the value of the `completion_time` field of the
`slide.slide` model from the website the traceback will be generated.

Steps to reproduce:
1. Install the `eLearning` module.
2. Go to the website and click on the courses menu.
3. Open the course which has channel type `document`.
   i.e: `Trees, Wood and Gardens`
4. Click on the `Edit` button.
5. Change the value of the `completion_time` field and save it.

Error: Traceback appears `ValueError: could not convert string to float:'00:00'`

when the user tries to change the value of `completion_time` field from the
website and save it at that time the value of `completion_time` gets the string
type and it is not converted to float. so, that the error is occurring.

This commit will solve the above issue by raising the UserError if the value of
`completion_time` field is not converted into float type.

Sentry-4288785196