[WIP] theme_default: configure a new page template
This is an example of how to configure a new page template from a given
theme.

Ultimately, this kind of example will be applied on `theme_test_custo`.

Currently, blocks need to be put in the manifest so that their primary
templates are generated - but the page generation itself is currently
skipped for themes, therefore the `sections` template needs to be
xpathed explicitly.

task-3381714