[FIX] web_editor: disable dropImageAsAttachment for mail composer
This commit implements a temporary solution to deactivate the conversion
of pasted or dropped images into attachments through a controller within
the web_editor module. This alteration enables the mail composer's
own mechanism to handle this process.

It is worth noting that the parent commit already addressed this issue by
modifying the view of the mail composer. This current commit serves to
rectify this behavior for databases that have yet to update the 'mail'
module.

task-3457541
opw-3448851