[FIX] web: close dialogs on command execution from command pallet
Before this commit:
Previously, when the command palette was opened and a command was being executed
with single or multiple dialog boxes behind it, only the command palette would
close. This left the dialog boxes open.

After this commit:
Now, when a command is being executed from the command palette, all the dialog
boxes and the command palette close simultaneously.

Task-3441132