[FIX] base:  allow only alphabetical characters as country code
If a country has a non-alphabetical country code, selecting a customer from that
country and clicking on the 'Send & Print' button to send the invoice,
will result in an error.

steps to produce:
- Install `l10n_es_edi_facturae`.
- Change the company to ES Company.
- Contacts > Configuration > Localization > Countries > Select a country
  (For eg: United States)
- Change the country code to any other character other than alphabetical and
  save them.
- Invoicing > Customer > Invoices > Select a customer having country whose
  country code you modified.
- Confirm the invoice, and then click on `send  & print`
- click on save button to save  the document layout, click on `send & print`
  again on Send wizard.

Error: `KeyError`

After applying this commit would not allow the user to change the country code
into something that is not alphabetical. `COUNTRY_CODE_MAP` is the dictionary
with only string keys and it does not have any numeric keys. Also country code
we have is Alpha-2 Code. In the standard code we do not have any numeric codes
so when the user enters a numeric code then they will face error while sending
the invoice.

sentry-4370088065