[REF] *: modifying tours to handle data-tooltip disparition
IMO we need to remove those test code snipets.
the problem is that the html element td[name='name'] does not contain information about the text inside the description while the focus
is still on the record. The only way we had to know what was inside the textarea before unfocusing was through the datatooltip
which doesn't exist anymore.