[IMP] web, website: handle `border-color` variations
This commit aims to harmonize default border-color values and to provide
a predictable result when transitioning between color-presets.

We assign a CSS variable to bootstrap `$border-color`in order
to let bootstrap components (and utility classes) align with the chosen
color-preset.

The color itself is achieved by fading-out the current text-color using
CSS 'mix-color()' function.

Since, at the time of writing, the function itself has a global support
of 81% (notably 'cause of IE), a fallback has been conveniently
provided.

[1] https://caniuse.com/mdn-css_types_color_color-mix

task-3097005