[IMP] website, *: handle `border-color` variations
*: website_sale

This commit aims to harmonize default border-color values and to provide
a predictable result when transitioning between color-presets.

We assign a CSS variable to bootstrap `$border-color` in order to let
bootstrap components (and utility classes) align with the chosen
color-preset.

The color itself is achieved by fading-out the current text-color using
CSS 'mix-color()' function.

Since, at the time of writing, the function itself has a global support
of 81% (mainly because of IE which we do not support but... see [1]), a
fallback has been conveniently provided.

Note: the right panel of the editor actually cannot understand the
result of the `mix-color` function either and will display a transparent
color instead. This will be fixed in a later PR.

[1]: https://caniuse.com/mdn-css_types_color_color-mix

task-3097005

Co-authored-by: Brieuc-brd <brd@odoo.com>