[FIX] website: make navbar content invisible when block invisible
Steps to reproduce:
 1. Go to the website
 2. Add website languages
 3. Drag Table Of Content
 4. Select snippet block
 5. Go to customize the block
 6. Select visibility conditionally - set visible/hide for languages.

Before the commit, selecting block visibility conditionally makes only
the block is invisible but not the navbar content is invisible.

In this commit, handling conditional visibility of the navbar when the
snippet block is invisible.

In this case, I gave it a shot and directly noticed it was working fine
with mobile visibility but not with conditional visibility. From there,
I found a commit [1] that is related, and that this code is modifying
actually (which would also lead you to this commit).
[1]: https://github.com/odoo/odoo/commit/291439da112b663f45541adcab93024e411bf2c1

task-3373943