[FIX] point_of_sale: same amount for percentage of child taxes
Before commit:
Any tax, having a group of included taxes with the same amount percentage, in
some cases the amount of included taxes are not the same.

After commit:
Any tax, having a group of included taxes with the same percentage, the amount
of child taxes is now the same.