[REF] {mrp,stock}_account: move prod accounts to stock_account
PR odoo/odoo#113973 originally added a production cost account for
locations of type Production Type. At the time it logically made sense
to add this to mrp_account. In retrospect, the Repair app also moves
products to/from production locations so this cost account + the
settings to use this feature should also be available when Repair is
installed. Unfortunately Repair does not depend on mrp and neither is
the reserve the case. Therefore we move this feature to the common
module of the two apps: stock_account.

It is no ideal since it will appear when neither repair nor mrp are
installed, but it is also not a huge deal since we already allow
locations to have `usage='production'` with only stock installed.

Task: 3471065