[IMP] web: Not use locale for float inputs parsing
Avoid faulty inputs caused by a weird parsing/detection mixup between thousand
and decimal separator. We must let go of the (wrong) assumption that people will
paste numbers into the format defined on their locale. We should not depend on
the locale for number parsing as much as possible.

The locale format can still be used for formatting of course, so this config has
its use - but parsing should be much more agnostic.

A new function has been added in this commit
"lenientParseNumber(value, options)"
that detects separators with regex and tries to interpret the correct value of
the number taking into account the separations.

task-3092583