[IMP] point_of_sale: add a "should be printed" method to the invoice
For some localization, an invoice is created automatically when creating
a pos_order. This leads to a PDF being donwloaded for every pos_order
done during a session which might be annoying for the cashier.

This commit adds a method that is meant to be overriden by those
localizations so that this invoice is not printed everytime.