[FIX] models: set origin module on field imd
For inherited fields, the module on the xmlid should reflect the module
in which the field comes from, and the module of the current context.

The order of installing modules can change the module on the xmlid
for the same field. which can be problematic when the module that
introduces the field is being uninstalled, if that module is not on
the xmlid of the inherited field, that field will not be removed.